Reland Fixed compiler warning C4267 'conversion from 'size_t' to 'type', possible loss of data' Additional warnings found with more testing and added C4267 warning disable only for angle_libpng
BUG=angleproject:1120

Reviewed-on: https://chromium-review.googlesource.com/293028
Tested-by: Geoff Lang <geofflang@chromium.org>

+++ src/common/BitSetIterator.h
@@ -84,7 +84,7 @@ BitSetIterator<N>::Iterator::Iterator(const std::bitset<N> &bits)
     }
     else
     {
-        mOffset = rx::roundUp(N, BitsPerWord);
+        mOffset = static_cast<unsigned long>(rx::roundUp(N, BitsPerWord));
     }
 }
