+++ src/tests/gl_tests/IncompleteTextureTest.cpp
@@ -134,8 +134,8 @@ TEST_P(IncompleteTextureTest, UpdateTexture)
     fillTextureData(redTextureData, 255, 0, 0, 255);
     for (size_t i = 0; i < 7; i++)
     {
-        glTexImage2D(GL_TEXTURE_2D, i, GL_RGBA, redTextureWidth >> i, redTextureHeight >> i, 0, GL_RGBA, GL_UNSIGNED_BYTE,
-                     &redTextureData[0]);
+        glTexImage2D(GL_TEXTURE_2D, static_cast<GLint>(i), GL_RGBA, redTextureWidth >> i,
+                     redTextureHeight >> i, 0, GL_RGBA, GL_UNSIGNED_BYTE, &redTextureData[0]);
     }
 
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
@@ -151,9 +151,9 @@ TEST_P(IncompleteTextureTest, UpdateTexture)
 
     for (size_t i = 0; i < 6; i++)
     {
-        glTexSubImage2D(GL_TEXTURE_2D, i, greenTextureWidth >> i, greenTextureHeight >> i,
-                        greenTextureWidth >> i, greenTextureHeight >> i, GL_RGBA, GL_UNSIGNED_BYTE,
-                        &greenTextureData[0]);
+        glTexSubImage2D(GL_TEXTURE_2D, static_cast<GLint>(i), greenTextureWidth >> i,
+                        greenTextureHeight >> i, greenTextureWidth >> i, greenTextureHeight >> i,
+                        GL_RGBA, GL_UNSIGNED_BYTE, &greenTextureData[0]);
     }
 
     drawQuad(mProgram, "position", 0.5f);
